== Timers == =General: *Updated more timer-related code. -follow-up to r175(@7ec89e3) -this also fixes some critical issues with aoe-skills. Please update ASAP. -Corrected some leftover 32bit typecasts. -thanks to @HerculesWS.

+++ src/char_sql/int_auction.c
@@ -226,7 +226,7 @@ void inter_auctions_fromsql(void)
 		}
 
 		if( auction->timestamp > now )
-			endtick = ((unsigned int)(auction->timestamp - now) * 1000) + tick;
+			endtick = ((int64)(auction->timestamp - now) * 1000) + tick;
 		else
 			endtick = tick + 10000; // 10 Second's to process ended auctions
